backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5
pg8000==1.29.8
python-dotenv==1.0.0
Werkzeug==2.3.7

# Scraper dependencies
selenium==4.15.2
beautifulsoup4==4.12.2
requests==2.31.0
python-dateutil==2.8.2
